refactor(translator): remove loot site translation logic and related files
The loot site translation logic was removed as it was no longer needed. This includes deleting the `lootSites.js` file and removing related functions in the `translator.js` controller. Additionally, the `addLootSiteTranslation` function was replaced with `addTranslationIfItDoesNotAlreadyExist` in `lootParsers.js` to handle translations more generically. Debug logs in `dropProcessor.js` were also removed to clean up the codebase.